restore NonEmptyList and NonEmptyVector
