+++ core/src/main/scala-2.13+/cats/data/ScalaVersionSpecificPackage.scala
@@ -4,9 +4,4 @@ package data
 
 abstract private[data] class ScalaVersionSpecificPackage {
   type NonEmptyLazyList[+A] = NonEmptyLazyList.Type[A]
-
-  type NonEmptyList[+A] = NonEmptyList.Type[A]
-
-  type NonEmptyVector[+A] = NonEmptyVector.Type[A]
-
 }
